+++ API/ServiceCollectionExtension.cs
@@ -1,7 +1,12 @@
-﻿using DAL;
+﻿using BLL;
+using DAL;
 using DAL.UnitOfWork;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
+using System.Text;
 
 namespace API
 {
@@ -48,6 +53,53 @@ namespace API
         public static IServiceCollection ConfigureRepositoryWrapper(this IServiceCollection services)
         {
             services.AddScoped<IUnitOfWork,UnitOfWork> ();
+
+            return services;
+        }
+
+
+        public static IServiceCollection ConfigureAuth(this IServiceCollection services, IConfiguration configuration)
+        {
+
+            // подключение к БД
+            var connectionStringUsers = configuration.GetConnectionString("DefaultConnection");
+            services.AddDbContext<UserDBContext>(options => options.UseMySQL(connectionStringUsers));
+            services.AddIdentity<IdentityUser, IdentityRole>().AddEntityFrameworkStores<UserDBContext>();
+
+            services.Configure<JWTSettings>(configuration.GetSection("JWTSettings"));
+
+            var secretKey = configuration.GetSection("JWTSettings:SecretKey").Value;
+            var issuer = configuration.GetSection("JWTSettings:Issuer").Value;
+            var audience = configuration.GetSection("JWTSettings:Audience").Value;
+            var singingKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(secretKey));
+
+            services.AddAuthentication(option =>
+            {
+                option.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+                option.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+                option.DefaultScheme = JwtBearerDefaults.AuthenticationScheme;
+            })
+
+            .AddJwtBearer(options =>
+            {
+                options.TokenValidationParameters = new TokenValidationParameters
+                {
+                    ValidateIssuer = true,
+                    ValidIssuer = issuer,
+                    ValidateAudience = true,
+                    ValidAudience = audience,
+                    ValidateLifetime = true,
+                    IssuerSigningKey = singingKey,
+                    ValidateIssuerSigningKey = true
+                };
+            });
+
+
+            // ограничиваем доступ к страницам
+            services.AddAuthorization(option =>
+            {
+                option.AddPolicy("OnlyAdmin", policyBuilder => policyBuilder.RequireClaim("Role", "1"));
+            });
 
             return services;
         }
