+++ DAL/Entities/User_Auth.cs
@@ -9,10 +9,8 @@ namespace DAL.Entities
 {
     public class User_Auth
     {
-        [Required]
         [DataType(DataType.EmailAddress)]
         public string Email { get; set; } = null!;
-        [Required]
         [DataType(DataType.Password)]
         public string Password { get; set; } = null!;
     }
